// src/app/store/companies/types.ts
export enum ACTION_TYPES {
    COMPANY_DETAILS_REQUEST = 'COMPANIES/COMPANY_DETAILS_REQUEST',
    COMPANY_DETAILS_REQUEST_SUCCESS = 'COMPANIES/COMPANY_DETAILS_REQUEST_SUCCESS',
    COMPANY_DETAILS_REQUEST_FAIL = 'COMPANIES/COMPANY_DETAILS_REQUEST_FAIL',
    SET_SELECTED_SERVICE_ID = 'COMPANIES/SET_SELECTED_SERVICE_ID',
    SLOTS_REQUEST = 'COMPANIES/SLOTS_REQUEST',
    SLOTS_REQUEST_SUCCESS = 'COMPANIES/SLOTS_REQUEST_SUCCESS',
    SLOTS_REQUEST_FAIL = 'COMPANIES/SLOTS_REQUEST_FAIL',
    SET_SELECTED_SLOT_ID = 'COMPANIES/SET_SELECTED_SLOT_ID',
    BOOK_SLOT_REQUEST = 'COMPANIES/BOOK_SLOT_REQUEST',
    BOOK_SLOT_REQUEST_SUCCESS = 'COMPANIES/BOOK_SLOT_REQUEST_SUCCESS',
    BOOK_SLOT_REQUEST_FAIL = 'COMPANIES/BOOK_SLOT_REQUEST_FAIL',
    SET_SELECTED_DATE = 'COMPANIES/SET_SELECTED_DATE',
    SELECTED_DATE_ADD_ONE = 'COMPANIES/SELECTED_DATE_ADD_ONE',
    SELECTED_DATE_SUBTRACT_ONE = 'COMPANIES/SELECTED_DATE_SUBTRACT_ONE',
    SELECTED_DATE_TODAY = 'COMPANIES/SELECTED_DATE_TODAY',
    CURRENT_CART_REQUEST = 'COMPANIES/CURRENT_CART_REQUEST',
    CURRENT_CART_REQUEST_SUCCESS = 'COMPANIES/CURRENT_CART_REQUEST_SUCCESS',
    CURRENT_CART_REQUEST_FAIL = 'COMPANIES/CURRENT_CART_REQUEST_FAIL',
    DELETE_APPOINTMENT_REQUEST = 'COMPANIES/DELETE_APPOINTMENT_REQUEST',
    DELETE_APPOINTMENT_REQUEST_SUCCESS = 'COMPANIES/DELETE_APPOINTMENT_REQUEST_SUCCESS',
    DELETE_APPOINTMENT_REQUEST_FAIL = 'COMPANIES/DELETE_APPOINTMENT_REQUEST_FAIL',
    SET_CURRENT_REQUEST = 'COMPANIES/SET_CURRENT_REQUEST',
}